Warnings added to Makefile. ALU WIP

+++ inc/file.h
@@ -2,7 +2,7 @@
 #define FILE_H
 #include <stddef.h>
 
-size_t file_len(const char name[static 1]);
-int    readFile(void *dst, const char name[static 1], unsigned max_len);
+long file_len(const char name[static 1]);
+int  readFile(void *dst, const char name[static 1], long max_len);
 
 #endif //  FILE_H
